// web/src/lib/api/generated/models/githubComMantonxViewraInternalApplicationMediaSubtitleTrackResponse.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * ViewRA Media Server API
 * Self-hosted media server for movies, TV shows, and music
 * OpenAPI spec version: 0.0.1
 */

export interface GithubComMantonxViewraInternalApplicationMediaSubtitleTrackResponse {
  codec?: string
  created_at?: string
  file_path?: string
  id?: number
  is_bitmap?: boolean
  is_commentary?: boolean
  is_default?: boolean
  is_forced?: boolean
  is_sdh?: boolean
  language?: string
  media_id?: number
  /** "embedded" or "external" */
  source_type?: string
  stream_index?: number
  title?: string
}